+++ databricks/config.go
@@ -10,6 +10,7 @@ import (
 	"github.com/innovationnorway/go-azure/auth"
 	"github.com/innovationnorway/go-databricks/clusters"
 	"github.com/innovationnorway/go-databricks/databricks"
+	"github.com/innovationnorway/go-databricks/dbfs"
 	"github.com/innovationnorway/go-databricks/groups"
 	"github.com/innovationnorway/go-databricks/workspace"
 	"github.com/innovationnorway/terraform-provider-databricks/version"
@@ -39,6 +40,7 @@ type AzureServicePrincipalConfig struct {
 
 type Meta struct {
 	Clusters    clusters.BaseClient
+	Dbfs        dbfs.BaseClient
 	Groups      groups.BaseClient
 	Workspace   workspace.BaseClient
 	StopContext context.Context
@@ -71,6 +73,9 @@ func (c *Config) createClients(baseURI string, authorizer autorest.Authorizer) (
 
 	meta.Clusters = clusters.NewWithBaseURI(baseURI)
 	configureClient(&meta.Clusters.Client, authorizer, c.terraformVersion)
+
+	meta.Dbfs = dbfs.NewWithBaseURI(baseURI)
+	configureClient(&meta.Dbfs.Client, authorizer, c.terraformVersion)
 
 	meta.Groups = groups.NewWithBaseURI(baseURI)
 	configureClient(&meta.Groups.Client, authorizer, c.terraformVersion)