+++ databricks/provider.go
@@ -86,6 +86,7 @@ func Provider() terraform.ResourceProvider {
 
 		ResourcesMap: map[string]*schema.Resource{
 			"databricks_cluster":          resourceDatabricksCluster(),
+			"databricks_dbfs_mkdirs":      resourceDatabricksDbfsMkdirs(),
 			"databricks_group":            resourceDatabricksGroup(),
 			"databricks_group_member":     resourceDatabricksGroupMember(),
 			"databricks_workspace_import": resourceDatabricksWorkspaceImport(),